requirements.txt
curl-cffi
bs4
python-dotenv
lxml
playwright